Added room integration

+++ app/src/main/java/com/korkalom/todolist/model/Task.kt
@@ -1,7 +1,7 @@
 package com.korkalom.todolist.model
 
 data class Task(
-    val taskID: Int = -1,
+    val taskID: Long = -1,
     val title: String = "No title",
     val description: String = "No description",
     val priority: Int = 1,
